Move audio config i/o into a separate class
