+++ bigdoors-core/src/main/java/nl/pim16aap2/bigdoors/audio/AudioConfigurator.java
@@ -1,10 +1,5 @@
 package nl.pim16aap2.bigdoors.audio;
 
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-import com.google.gson.JsonElement;
-import com.google.gson.JsonObject;
-import com.google.gson.JsonParser;
 import lombok.extern.flogger.Flogger;
 import nl.pim16aap2.bigdoors.api.debugging.DebuggableRegistry;
 import nl.pim16aap2.bigdoors.api.debugging.IDebuggable;
@@ -16,18 +11,10 @@ import nl.pim16aap2.bigdoors.managers.DoorTypeManager;
 import org.jetbrains.annotations.Nullable;
 
 import javax.inject.Inject;
-import javax.inject.Named;
-import java.io.IOException;
-import java.io.Reader;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.StandardOpenOption;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.Map;
-import java.util.logging.Level;
 
 /**
  * Represents a class used to read and access the audio configuration.
@@ -37,26 +24,20 @@ import java.util.logging.Level;
 @Flogger
 public final class AudioConfigurator implements IRestartable, IDebuggable
 {
+    public static final String KEY_DEFAULT = "DEFAULT";
     private static final AudioSet EMPTY_AUDIO_SET = new AudioSet(null, null);
-    private static final String KEY_DEFAULT = "DEFAULT";
 
-    private final Path file;
+    private final AudioConfigIO audioConfigIO;
     private final DoorTypeManager doorTypeManager;
     private final Map<DoorType, AudioSet> audioMap = new HashMap<>();
 
-    private final Gson gson = new GsonBuilder()
-        .serializeNulls()
-        .setPrettyPrinting()
-        .registerTypeAdapter(AudioSet.class, new AudioSet.Deserializer())
-        .registerTypeAdapter(AudioDescription.class, new AudioDescription.Deserializer())
-        .create();
 
     @Inject//
     AudioConfigurator(
-        @Named("pluginBaseDirectory") Path baseDir, RestartableHolder restartableHolder,
+        AudioConfigIO audioConfigIO, RestartableHolder restartableHolder,
         DebuggableRegistry debuggableRegistry, DoorTypeManager doorTypeManager)
     {
-        file = baseDir.resolve("audio_config.json");
+        this.audioConfigIO = audioConfigIO;
         this.doorTypeManager = doorTypeManager;
 
         restartableHolder.registerRestartable(this);
@@ -81,18 +62,26 @@ public final class AudioConfigurator implements IRestartable, IDebuggable
 
     private void processAudioConfig()
     {
+        final Map<DoorType, @Nullable AudioSet> defaults = getDefaults();
+        final Map<String, @Nullable AudioSet> parsed = audioConfigIO.readConfig();
+        final @Nullable AudioSet defaultAudioSet = parsed.get(KEY_DEFAULT);
+
+        final Map<DoorType, @Nullable AudioSet> merged = mergeMaps(parsed, defaults, defaultAudioSet);
+
+        audioConfigIO.writeConfig(merged, defaultAudioSet);
+
+        merged.forEach((key, val) -> audioMap.put(key, val == null ? EMPTY_AUDIO_SET : val));
+    }
+
+    /**
+     * @return The default audio set mappings for each door type as specified by the door type itself.
+     */
+    private Map<DoorType, @Nullable AudioSet> getDefaults()
+    {
         final Collection<DoorType> enabledDoorTypes = doorTypeManager.getEnabledDoorTypes();
         final Map<DoorType, @Nullable AudioSet> defaultMap = new HashMap<>(enabledDoorTypes.size());
         doorTypeManager.getEnabledDoorTypes().forEach(type -> defaultMap.put(type, type.getAudioSet()));
-
-        final Map<String, @Nullable AudioSet> parsed = readConfig();
-        final @Nullable AudioSet defaultAudioSet = parsed.get(KEY_DEFAULT);
-
-        final Map<DoorType, @Nullable AudioSet> merged = mergeMaps(parsed, defaultMap, defaultAudioSet);
-
-        writeConfig(merged, defaultAudioSet);
-
-        merged.forEach((key, val) -> audioMap.put(key, val == null ? EMPTY_AUDIO_SET : val));
+        return defaultMap;
     }
 
     private Map<DoorType, @Nullable AudioSet> mergeMaps(
@@ -114,69 +103,6 @@ public final class AudioConfigurator implements IRestartable, IDebuggable
         return merged;
     }
 
-    private Map<String, @Nullable AudioSet> readConfig()
-    {
-        if (!Files.isRegularFile(file))
-            return Collections.emptyMap();
-
-        // Older versions of gson don't have the new method yet, so this avoids
-        // breaking on older versions.
-        @SuppressWarnings("deprecation")//
-        final JsonParser jsonParser = new JsonParser();
-        try
-        {
-            final Reader reader = Files.newBufferedReader(file);
-            @SuppressWarnings("deprecation")//
-            final JsonObject base = jsonParser.parse(reader).getAsJsonObject();
-            return readConfig(base);
-        }
-        catch (Exception | AssertionError e)
-        {
-            log.at(Level.SEVERE).withCause(e).log("Failed to read audio config!");
-        }
-        return Collections.emptyMap();
-    }
-
-    private Map<String, @Nullable AudioSet> readConfig(JsonObject base)
-    {
-        final Map<String, @Nullable AudioSet> parsed = new LinkedHashMap<>();
-        for (final Map.Entry<String, JsonElement> entry : base.getAsJsonObject().entrySet())
-        {
-            final String name = entry.getKey();
-            final @Nullable AudioSet audioSet;
-            if ("\"null\"".equals(entry.getValue().toString()))
-                audioSet = null;
-            else
-                audioSet = gson.fromJson(entry.getValue(), AudioSet.class);
-            parsed.put(name, audioSet);
-        }
-        return parsed;
-    }
-
-    private void writeConfig(Map<DoorType, @Nullable AudioSet> defaultMap, @Nullable AudioSet defaultAudioSet)
-    {
-        final JsonObject base = new JsonObject();
-        appendToJsonObject(gson, base, KEY_DEFAULT, defaultAudioSet);
-        defaultMap.forEach((key, val) -> appendToJsonObject(gson, base, key.getSimpleName(), val));
-
-        try
-        {
-            Files.writeString(file, gson.toJson(base), StandardOpenOption.TRUNCATE_EXISTING, StandardOpenOption.CREATE);
-        }
-        catch (IOException e)
-        {
-            log.at(Level.SEVERE).withCause(e).log("Failed to write audio config!");
-        }
-    }
-
-    private void appendToJsonObject(Gson gson, JsonObject jsonObject, String typeName, @Nullable AudioSet audioSet)
-    {
-        if (audioSet == null || audioSet.isEmpty())
-            jsonObject.addProperty(typeName, gson.toJson(null));
-        else
-            jsonObject.add(typeName, gson.toJsonTree(audioSet));
-    }
-
     @Override
     public synchronized void initialize()
     {
@@ -193,8 +119,7 @@ public final class AudioConfigurator implements IRestartable, IDebuggable
     public String getDebugInformation()
     {
         final StringBuilder sb = new StringBuilder();
-        sb.append("AudioConfigReader File: '").append(file).append("'\n")
-          .append("AudioSets:\n");
+        sb.append("AudioSets:\n");
         audioMap.forEach((key, val) -> sb.append("  Type: ").append(key).append(", Audio: ").append(val).append('\n'));
         return sb.toString();
     }
